@@ -1,6 +1,6 @@
 import SelectionUtils from '../selection';
 
-declare var Module: any;
+import Module from '../__module';
 
 export default class DragNDrop extends Module {
 
@@ -67,7 +67,7 @@ export default class DragNDrop extends Module {
      * If drop target (error will be thrown) is not part of the Block, set last Block as current.
      */
     try {
-      BlockManager.setCurrentBlockByChildNode(dropEvent.target, 'end');
+      BlockManager.setCurrentBlockByChildNode(dropEvent.target as Node, 'end');
     } catch (e) {
       BlockManager.setCurrentBlockByChildNode(BlockManager.lastBlock.holder, 'end');
     }
