@@ -7,7 +7,7 @@
 
 import IEditorConfig from '../interfaces/editor-config';
 
-declare const Module: any;
+import Module from '../__module';
 declare const _: any;
 
 export default class ModificationsObserver extends Module {
@@ -23,7 +23,7 @@ export default class ModificationsObserver extends Module {
    * @type {Function}
    */
   private mutationDebouncer = _.debounce( () => {
-    this.config.onChange.call();
+    this.config.onChange();
   }, ModificationsObserver.DebounceTimer);
 
   /**
