@@ -4,7 +4,7 @@
  * @property {Function} function - function's that must be called asynchronically
  */
 interface ChainData {
-  data: any;
+  data?: any;
   function: (...args: any[]) => any;
 }
 
@@ -45,7 +45,7 @@ export default class Util {
    * Returns basic keycodes as constants
    * @return {{}}
    */
-  static get keyCodes(): object {
+  static get keyCodes() {
     return {
       BACKSPACE: 8,
       TAB: 9,
@@ -73,7 +73,11 @@ export default class Util {
    *
    * @return {Promise}
    */
-  public static async sequence(chains: ChainData[], success = () => {}, fallback = () => {}): Promise<void> {
+  public static async sequence(
+    chains: ChainData[],
+    success: (data: any) => void = () => {},
+    fallback: (data: any) => void = () => {},
+  ): Promise<void> {
     /**
      * Decorator
      *
